requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
aioconsole>=0.6.0
